+++ packages/visual-grid-client/src/sdk/getUserAgentForBrowser.js
@@ -1,3 +1,5 @@
+const {getBrowserKeyForUserAgent} = require('@applitools/eyes-sdk-core/shared')
+
 const userAgent = {
   IE:
     'Mozilla/5.0 (Windows NT 10.0; WOW64; Trident/7.0; .NET4.0C; .NET4.0E; .NET CLR 2.0.50727; .NET CLR 3.0.30729; .NET CLR 3.5.30729; rv:11.0) like Gecko',
@@ -14,12 +16,7 @@ const userAgent = {
 
 function getUserAgentForBrowser(browserName) {
   if (!browserName) return ''
-  if (browserName === 'ie10' || browserName === 'ie11') return userAgent.IE
-  else if (browserName.includes('chrome')) return userAgent.Chrome
-  else if (browserName.includes('firefox')) return userAgent.Firefox
-  else if (browserName.includes('safari')) return userAgent.Safari
-  else if (browserName.includes('edgechromium')) return userAgent.Edgechromium
-  else if (browserName.includes('edge')) return userAgent.Edge
+  return userAgent[getBrowserKeyForUserAgent(browserName)]
 }
 
 module.exports = {getUserAgentForBrowser, userAgent}